# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CODD",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/aparts/": {
      "get": {
        "tags": [
          "aparts"
        ],
        "summary": "Read Aparts",
        "operationId": "read_aparts",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApartList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/aparts/{id}": {
      "get": {
        "tags": [
          "aparts"
        ],
        "summary": "Read Apart",
        "operationId": "read_apart",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_rep",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id Rep"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Apart"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/aparts/create": {
      "post": {
        "tags": [
          "aparts"
        ],
        "summary": "Create Apart",
        "operationId": "create_apart",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApartBase"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Apart"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/aparts/checked": {
      "put": {
        "tags": [
          "aparts"
        ],
        "summary": "Checked Apart",
        "operationId": "checked_apart",
        "parameters": [
          {
            "name": "id_rep",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id Rep"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Apart"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/applogs/{id}": {
      "get": {
        "tags": [
          "applogs"
        ],
        "summary": "Get Log App",
        "operationId": "get_log_app",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_log",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id Log"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogsApp"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/applogs/": {
      "get": {
        "tags": [
          "applogs"
        ],
        "summary": "Get Logs App",
        "operationId": "get_logs_app",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "obj",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Obj"
            }
          },
          {
            "name": "user_uid",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Uid"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogsList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/applogs/create": {
      "post": {
        "tags": [
          "applogs"
        ],
        "summary": "Create Log App",
        "operationId": "create_log_app",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogsAppBase"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogsApp"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/applogs/csv/": {
      "get": {
        "tags": [
          "applogs"
        ],
        "summary": "Export Logs To Csv",
        "operationId": "export_logs_to_csv",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "object",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Object"
            }
          },
          {
            "name": "user_uid",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Uid"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bulvars/create": {
      "post": {
        "tags": [
          "bulvars"
        ],
        "summary": "Create Bulvar",
        "operationId": "create_bulvar",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulvarBase"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bulvar"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bulvars/": {
      "get": {
        "tags": [
          "bulvars"
        ],
        "summary": "Read Bulvars",
        "operationId": "read_bulvars",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulvarList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bulvars/depo": {
      "get": {
        "tags": [
          "bulvars"
        ],
        "summary": "Depo Bulvars",
        "operationId": "depo_bulvars",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "depo",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Depo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulvarList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Apart": {
        "properties": {
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          },
          "incident": {
            "type": "string",
            "title": "Incident"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "checked": {
            "type": "boolean",
            "title": "Checked",
            "default": false
          },
          "last_date": {
            "type": "string",
            "format": "date",
            "title": "Last Date"
          },
          "last_time": {
            "type": "string",
            "format": "time",
            "title": "Last Time"
          },
          "id_rep": {
            "type": "integer",
            "title": "Id Rep"
          }
        },
        "type": "object",
        "required": [
          "latitude",
          "longitude",
          "incident"
        ],
        "title": "Apart"
      },
      "ApartBase": {
        "properties": {
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          },
          "incident": {
            "type": "string",
            "title": "Incident"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "checked": {
            "type": "boolean",
            "title": "Checked",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "latitude",
          "longitude",
          "incident"
        ],
        "title": "ApartBase"
      },
      "ApartList": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Apart"
            },
            "type": "array",
            "title": "Data"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "data",
          "count"
        ],
        "title": "ApartList"
      },
...
      "Bulvar": {
        "properties": {
          "nfs": {
            "type": "number",
            "title": "Nfs"
          },
          "depo": {
            "type": "string",
            "title": "Depo"
          },
          "lt_loc_start": {
            "type": "number",
            "title": "Lt Loc Start"
          },
          "ln_loc_start": {
            "type": "number",
            "title": "Ln Loc Start"
          },
          "lt_loc_end": {
            "type": "number",
            "title": "Lt Loc End"
          },
          "ln_loc_end": {
            "type": "number",
            "title": "Ln Loc End"
          },
          "id_bul": {
            "type": "integer",
            "title": "Id Bul"
          },
          "last_date": {
            "type": "string",
            "format": "date",
            "title": "Last Date"
          },
          "last_time": {
            "type": "string",
            "format": "time",
            "title": "Last Time"
          }
        },
        "type": "object",
        "required": [
          "nfs",
          "depo",
          "lt_loc_start",
          "ln_loc_start",
          "lt_loc_end",
          "ln_loc_end"
        ],
        "title": "Bulvar"
      },
      "BulvarBase": {
        "properties": {
          "nfs": {
            "type": "number",
            "title": "Nfs"
          },
          "depo": {
            "type": "string",
            "title": "Depo"
          },
          "lt_loc_start": {
            "type": "number",
            "title": "Lt Loc Start"
          },
          "ln_loc_start": {
            "type": "number",
            "title": "Ln Loc Start"
          },
          "lt_loc_end": {
            "type": "number",
            "title": "Lt Loc End"
          },
          "ln_loc_end": {
            "type": "number",
            "title": "Ln Loc End"
          }
        },
        "type": "object",
        "required": [
          "nfs",
          "depo",
          "lt_loc_start",
          "ln_loc_start",
          "lt_loc_end",
          "ln_loc_end"
        ],
        "title": "BulvarBase"
      },
      "BulvarList": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Bulvar"
            },
            "type": "array",
            "title": "Data"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "data",
          "count"
        ],
        "title": "BulvarList"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LogsApp": {
        "properties": {
          "object": {
            "type": "string",
            "title": "Object"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "id_log": {
            "type": "integer",
            "title": "Id Log"
          },
          "last_date": {
            "type": "string",
            "format": "date",
            "title": "Last Date"
          },
          "last_time": {
            "type": "string",
            "format": "time",
            "title": "Last Time"
          }
        },
        "type": "object",
        "required": [
          "object",
          "action"
        ],
        "title": "LogsApp"
      },
      "LogsAppBase": {
        "properties": {
          "object": {
            "type": "string",
            "title": "Object"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "object",
          "action"
        ],
        "title": "LogsAppBase"
      },
      "LogsList": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/LogsApp"
            },
            "type": "array",
            "title": "Data"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "data",
          "count"
        ],
        "title": "LogsList"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        }
      }
    }
  }
}